* fix signed/unsigned comparison in longest_common_prefix
 * Fix crash when $postponed is on another server.
 * bcache: cast to avoid implicit signed/unsigned comparison in bcache_path
 * drop unused flags argument from imap_access

+++ bcache.c
@@ -74,7 +74,7 @@ static int bcache_path(struct Account *account, const char *mailbox, char *dst, 
 
   mutt_debug(3, "bcache_path: rc: %d, path: '%s'\n", len, dst);
 
-  if (len < 0 || len >= dstlen - 1)
+  if (len < 0 || (size_t) len >= dstlen - 1)
     return -1;
 
   mutt_debug(3, "bcache_path: directory: '%s'\n", dst);
